Cleaning an reorganizing code
