+++ Source/Extensions/Blazorise.Charts.Annotation/ChartAnnotation.razor.cs
@@ -13,22 +13,34 @@ namespace Blazorise.Charts.Annotation;
 /// <typeparam name="TItem">Data point type.</typeparam>
 public partial class ChartAnnotation<TItem> : ChartPlugin<TItem, JSChartAnnotationModule>
 {
-    
     #region Methods
-    protected override JSChartAnnotationModule GetNewJsModule() => new( JSRuntime, VersionProvider, BlazoriseOptions );
 
-    protected override async Task InitializePluginByJsModule() => await JSModule.AddAnnotationOptions( ParentChart.ElementId, Options );
+    /// <inheritdoc/>
+    protected override JSChartAnnotationModule GetNewJsModule()
+        => new( JSRuntime, VersionProvider, BlazoriseOptions );
 
-    protected override bool InitPluginInParameterSet( ParameterView parameters ) 
+    /// <inheritdoc/>
+    protected override async Task InitializePluginByJsModule()
+        => await JSModule.AddAnnotationOptions( ParentChart.ElementId, Options );
+
+    /// <inheritdoc/>
+    protected override bool InitPluginInParameterSet( ParameterView parameters )
         => parameters.TryGetValue<Dictionary<string, ChartAnnotationOptions>>( nameof( Options ), out var paramOptions ) && !Options.IsEqual( paramOptions );
 
     #endregion
-    protected override JSChartAnnotationModule JSModule { get;  set; }
-    
+
+    #region Properties
+
+    /// <inheritdoc/>
+    protected override string Name => "DataAnnotation";
+
+    /// <inheritdoc/>
+    protected override JSChartAnnotationModule JSModule { get; set; }
+
     /// <summary>
     /// Defines the options for an annotation.
     /// </summary>
     [Parameter] public Dictionary<string, ChartAnnotationOptions> Options { get; set; }
 
-    protected override string Name => "DataAnnotation";
+    #endregion
 }
