+++ Source/Extensions/Blazorise.Charts.Zoom/ChartZoom.razor.cs
@@ -12,22 +12,28 @@ namespace Blazorise.Charts.Zoom;
 /// <typeparam name="TItem">Data point type.</typeparam>
 public partial class ChartZoom<TItem> : ChartPlugin<TItem, JSChartZoomModule>
 {
-
     #region Methods
 
-    protected override JSChartZoomModule GetNewJsModule() => new( JSRuntime, VersionProvider, BlazoriseOptions );
+    /// <inheritdoc/>
+    protected override JSChartZoomModule GetNewJsModule()
+        => new( JSRuntime, VersionProvider, BlazoriseOptions );
 
-    protected override async Task InitializePluginByJsModule() => await JSModule.AddZoom( ParentChart.ElementId, Options );
+    /// <inheritdoc/>
+    protected override async Task InitializePluginByJsModule()
+        => await JSModule.AddZoom( ParentChart.ElementId, Options );
 
-    protected override bool InitPluginInParameterSet( ParameterView parameters ) 
+    /// <inheritdoc/>
+    protected override bool InitPluginInParameterSet( ParameterView parameters )
         => parameters.TryGetValue<ChartZoomPluginOptions>( nameof( Options ), out var paramOptions ) && !Options.IsEqual( paramOptions );
 
     #endregion
 
     #region Properties
-    
+
+    /// <inheritdoc/>
     protected override string Name => "Zoom";
 
+    /// <inheritdoc/>
     protected override JSChartZoomModule JSModule { get; set; }
 
     /// <summary>
